server/requirements.txt
bcrypt            1.6.3
Flask-Bcrypt      1.0.1
Flask-Migrate     4.0.5
Flask-MySQL       1.5.2
Flask-MySQLdb     2.0.0
Flask-SQLAlchemy  3.1.1
pytest            7.4.2
flask_cors
flask_mail
